[ cpp ] Fix Native[Double|Long|Short]

+++ core/jni/cpp_interop.cpp
@@ -70,9 +70,9 @@ JavaCritical_org_ice1000_jimgui_Native ## JavaName ## _deallocateNativeObject0(j
 
 NativeImpls(Int, jint, int, 0)
 NativeImpls(Float, jfloat, float, 0)
-NativeImpls(Double, jdouble, float, 0)
-NativeImpls(Short, jshort, float, 0)
-NativeImpls(Long, jlong, float, 0)
+NativeImpls(Double, jdouble, double, 0)
+NativeImpls(Short, jshort, short, 0)
+NativeImpls(Long, jlong, long, 0)
 
 #undef NativeImpls
